requirements.txt
# Core Slack Bot Dependencies
slack-bolt>=1.18.0
slack-sdk>=3.21.0

# Async Support
asyncio-mqtt>=0.11.1

# Database and State Management
supabase>=1.0.4
redis>=4.5.4
psycopg2-binary>=2.9.6

# AI/LLM Dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.1.0
langchain-core>=0.1.0
openai>=1.12.0

# Configuration Management
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Development Tools
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Logging and Monitoring
structlog>=23.1.0
prometheus-client>=0.17.0

# HTTP Client for API calls
httpx>=0.24.1
aiohttp>=3.8.5
urllib3>=1.26.0,<2.0  # Pin to v1.x for LibreSSL compatibility on macOS

# Web Search Tool Dependencies
beautifulsoup4>=4.12.0

# Utilities
python-dateutil>=2.8.2
jsonschema>=4.19.0 